Feat/run test on push ti=o dev (#7)
* ✨ feat: Add GitHub Workflow to run unit tests

The commit adds a GitHub workflow for running unit tests on pull requests. It uses Ubuntu latest as the runner and includes steps to checkout, set up Node.js version 18, and run tests using pnpm.

* ✨ feat: limit unit tests to "dev" branch in GitHub workflow

This commit modifies the ".github/workflows/run-unit-tests.yml" file to run unit tests only on the "dev" branch of the repository when a pull request is made. The change improves the efficiency of the testing process and ensures that only code changes from the development branch are tested.

* 🔧 chore: update test runner from pnpm to npm

* ✨ feat: Add workflow to install dependencies and run unit tests

This commit adds a GitHub Actions workflow that installs project dependencies with `npm ci` and runs the unit tests with `npm test`. The workflow is triggered on every push to the repository.

* ✨ feat: Install and use pnpm for package installation

This commit updates the GitHub Actions workflow to install pnpm globally using npm and use it as the package manager for installing dependencies. This change aims to improve package installation performance by taking advantage of pnpm's caching capabilities.

* ✨ feat: Update pnpm version in run-unit-tests.yml

Updates the installation of pnpm in the GitHub Actions workflow to version 7.18.2.

* ✨ feat: Add vitest configuration file

This commit adds a new file `vitest.config.ts` with the initial configuration for Vitest. The configuration is based on the existing Vite configuration (`vite.config.ts`) and includes additional settings for testing, such as global variables and environment match globs.

* ✨ feat: Add branch restriction for unit test workflows

This commit adds a branch restriction to the 'on' section of the '.github/workflows/run-unit-tests.yml' file. Unit tests will now only run when changes are pushed or pull requests are made to the 'dev' branch, ensuring that development work is properly tested before merging into the main branch.