Enhance/add styles btn behaviour (#13)
* remove global unused css styles

* install sass

* ✨ feat: Add new colors and border radius to tailwind.config.js

The commit adds new color options to the Tailwind configuration, including mintgreen and netural with different shades. Additionally, it extends the configuration to include a new border radius option of xs.

* ✨ feat: Update CustomIcon font sizes and add font weight

This commit updates the font-size classes in CustomIcon.module.css for small, medium, and large icons. It also adds a font-weight of bold to the CustomIcon component in index.tsx.

* ✨ feat: Add disabled state to Button component

This commit adds a `disabled` boolean prop to the `IButton` interface and implements it in the `Button` component. Additionally, styles for the disabled state have been added to the `.filled` button class in `Button.module.scss`. The `Default` story in `Button.stories.tsx` has also been updated to include the new `disabled` prop.