✨ feat: Add varient prop to Typography component and implement its functionality (#18)
This commit adds a new optional `varient` prop to the `Typography` component that allows for different styles to be applied to typography elements. The possible values for this prop are `"sectionTitle"`, `"subSectionTitle"`, `"cardTitle"`, and `"cardTag"`. Additionally, the implementation of this new prop has been added to the `Typography` component, with corresponding CSS classes defined in the `Typography.module.css` file. Finally, the `tailwind.config.ts` file has been updated to include some new box shadow options.